caloriecounter improvements: dummy data loader, adding new meals

+++ week-09/caloriecounter/src/main/java/com/greenfoxacademy/caloriecounter/domain/MealRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MealRepository extends CrudRepository<Meal, Long> {
+public interface MealRepository extends CrudRepository<Meal, Integer> {
 }
